+++ lib/injectable.rb
@@ -5,6 +5,7 @@ require 'injectable/dependencies_proxy'
 require 'injectable/dependency'
 require 'injectable/instance_methods'
 require 'injectable/missing_dependencies_exception'
+require 'injectable/method_already_exists_exception'
 
 # Convert your class into an injectable service
 #