+++ lib/injectable/method_already_exists_exception.rb
@@ -1,0 +1,4 @@
+module Injectable
+  class MethodAlreadyExistsException < RuntimeError
+  end
+end